create base tank class and controller logic.
HorizontalMoveController and SpinShootController are basic examples of how we might make different controllers.

Logic for creating a new Tank in the TestState class isn't exactly how we'll want to do it long-term. Better to set up a factory pattern kind of system for instantiating the kinds of tanks we want.